[Feat] Set FIrestore authenticationCode Alert
